pyproject.toml
[project]
name = "data-analytics-retrieval"
version = "0.1.0"
description = "CLI tool for retrieving and filtering GA4 events using cross-encoder scoring"
requires-python = ">=3.10"
dependencies = [
    "openpyxl>=3.1",
    "cohere>=5.0",
    "pinecone>=5.0",
    "rank-bm25>=0.2",
    "google-genai>=1.0",
    "python-dotenv>=1.0",
    "numpy>=1.24",
    "streamlit>=1.30",
    "nltk>=3.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
